dev-requirements.txt
coverage
pytest
pytest-flask
psycopg2
testing.postgresql
